Add files via upload
Task 1 
This task requires estimating several models to identify the best model to predict default based on the available data.
Create a new variable from trainData called “y” which takes the value = 1 if the column “loan status” has the value “Charged Off” and 0 otherwise. All other variables provided to you other than the loan status are features or “predictors”. Consider whether you would like to transform your variables; for example, consider converting some of the categorical variables into a continuous variable. Identify also the 10 most correlated and the 10 least correlated variables.
1.1
Linear Regression Model
Fit a linear regression model to the trainData, with y as the outcome variable, with all the predictors.
(a) What is the Mean Squared Error for the training data?
(b) What is the Mean Squared Error for the testing data?
(c) How do the MSEs change when the analysis is limited to the 10 most correlated variables?
1.2 Ridge Regression Model
Fit a ridge regression model to the trainData, with y as the outcome variable, with the predictors. Explore all values of hyperparameter λ (lambda) ranging from 0.01 to 100 with an increment of 0.01.
(a) What is the Mean Squared Error for the “best” model of this class for the training data?
(b) What is the Mean Squared Error for the “best” model of this class for the test data?
(c) How do the MSEs change when the analysis is limited to the 10 most correlated variables?
1.3 Lasso Regresion Model
Fit a Lasso regression model to the trainData, with y as the outcome variable, with the predictors. Explore the same values of hyperparameter λ (lambda) as for the ridge regression.
(a) What is the Mean Squared Error for the “best” model of this class for the training data?
(b) What is the Mean Squared Error for the “best” model of this class for the test data?
(c) How do the MSEs change when the analysis is limited to the 10 most correlated variables?
1.4 Random Forest
Fit a randomForest to the trainData, with y as the outcome variable, with the predictors. Explore and fit the best model of this class. Please explain any estimation assumptions you take.
(a) What is the Mean Squared Error for the “best” model of this class for the training data?
(b) What is the Mean Squared Error for the “best” model of this class for the test data?
(c) How do the MSEs change when the analysis is limited to the 10 most correlated variables?
1.5 Neural Network
Select a Neural Network model and follow the same process as in the previous models.
(a) What is the Accuracy for the model for the training data?
(b) What is the Accuracy for the model for the testing data?
(c) Explain why you chose the particular Neural Network model